Extract Habits.time_series_check_in_data
And add a test for it

I’m not sure this ultimately belongs in here, but it feels like a better place than the schema.
